Move OZK tests and programs to same files
This way, we don't have to make the test-structs and functions visible
to the entire crate.
