+++ src/tests_and_benchmarks/ozk/programs.rs
@@ -2,3 +2,4 @@ pub mod bfe_add;
 pub mod removal_record_integrity_partial;
 pub mod simple_map_on_bfe;
 pub mod simple_struct;
+pub mod struct_with_vec;
