--- src/tests_and_benchmarks/ozk/tests.rs
@@ -1,4 +1,0 @@
-pub mod bfe_add;
-pub mod removal_record_integrity_partial;
-pub mod simple_map_on_bfe;
-pub mod simple_struct;
